// src/enums/book_category.enum.ts
export enum BookCategoryEnum {
  CULTURA = 'Cultura',
  ARTE = 'Arte',
  LENGUA = 'Lengua',
  PALENQUE = 'Palenque',
  RELIGION = 'Religion',
  LITERATURA = 'Literatura',
  TRADICION_ORAL = 'Tradicion Oral',
  GASTRONOMIA = 'Gastronomia',
  DOCUMENTALES = 'Documentales',
  PELICULAS = 'Peliculas',
  CORTROMETRAJES = 'Contrometrajes',
}