// src/constants/books.constant.ts
import { BookCategoryEnum } from "../enums/book_category.enum"

export interface IBook {
  name: string
  image: string
  url: string
  type: BookCategoryEnum
}

export const BOOKS: Array<IBook> = [
  {
    name: 'La lengua palenquera juvenil: contacto y conflicto de estructuras gramaticales',
    image: 'https://imgv2-2-f.scribdassets.com/img/document/317454296/original/ff01ebc0ba/1577309335?v=1',
    url: 'http://www.personal.psu.edu/jml34/Palenquero-UniverSOS.pdf',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Patterns of variable tense-aspect marking in Palenquero creole',
    image: 'https://0.academia-photos.com/attachment_thumbnails/34729044/mini_magick20190321-1974-1hwkjv0.png?1553171015',
    url: 'https://www.academia.edu/8322181/Patterns_of_variable_tense-aspect_marking_in_Palenquero_creole',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'EL LÉXICO NEGRO-AFRICANO DE SAN BASILIO DE PALENQUE',
    image: 'https://imgv2-1-f.scribdassets.com/img/document/138350602/original/7dbbbce11b/1591682197?v=1',
    url: 'https://cvc.cervantes.es/lengua/thesaurus/pdf/39/TH_39_123_100_0.pdf',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Comportamientos y actitudes lingüisticas de los hablantes bilingües de la comunidad palenquera',
    image: 'https://revistas.unimagdalena.edu.co/public/journals/4/cover_issue_63_es_ES.jpg',
    url: 'https://revistas.unimagdalena.edu.co/index.php/jangwapana/article/view/581/543',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Criollistica Afrocolombiana',
    image: 'https://image.isu.pub/160725200129-43fa6649867acf041686dcacf601f411/jpg/page_1.jpg',
    url: 'https://studylib.es/doc/1919279/download-this-file--marianne-dieck-criollistica-palenquer',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'San Basilio De Palenque, restacatando su lengua: La iniciativa comunitaria como modelo de gestión en el rescate de la lengua palenquera',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.png',
    url: 'https://studylib.es/doc/1919279/download-this-file--marianne-dieck-criollistica-palenquer',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'La expresión de la reflexividad en palenquero',
    image: 'https://0.academia-photos.com/attachment_thumbnails/37247127/mini_magick20190304-12149-1r05nel.png?1551710302',
    url: 'https://www.academia.edu/11850980/La_expresi%C3%B3n_de_la_reflexividad_en_palenquero',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: '"Abrakabraka",   "Suebbesuebbe" y otras voces Palenqueras: sus orígenes e importancia para el estudio de dialectos AfrohispanoCaribeños',
    image: 'https://docplayer.es/docs-images/94/118755720/images/18-0.jpg',
    url: 'http://www.humanas.unal.edu.co/colantropos/files/6814/7941/4914/afrakabraka.pdf',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Territorio palenquero, lengua como identidad cultural : situación de la lengua palenquera en la población juvenil de San Basilio de Palenque',
    image: 'https://repository.javeriana.edu.co/bitstream/handle/10554/34050/Tesis%20palenquero%20entrega%20final.pdf.jpg?sequence=2&isAllowed=y',
    url: 'https://repository.javeriana.edu.co/handle/10554/34050',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Los criollos con negación post-oracional: Estudio comparativo',
    image: 'https://0.academia-photos.com/attachment_thumbnails/42141053/mini_magick20180819-22050-1r6civb.png?1534673460',
    url: 'https://www.academia.edu/5868675/Los_criollos_con_negaci%C3%B3n_post_oracional_Estudio_comparativo',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'La partícula focal "jue" (< español "fue") en el criollo palenquero: ¿gramaticalización y/o sustrato?',
    image: 'https://0.academia-photos.com/attachment_thumbnails/55583560/mini_magick20180818-18906-p7fcfp.png?1534608837',
    url: 'https://www.academia.edu/31794241/La_part%C3%ADcula_focal_jue_espa%C3%B1ol_fue_en_el_criollo_palenquero_gramaticalizaci%C3%B3n_y_o_sustrato',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'La lengua de palenque: avances en la investigación de su estructura gramatica',
    image: 'https://0.academia-photos.com/attachment_thumbnails/35813379/mini_magick20190311-14904-18o7fox.png?1552347742',
    url: 'https://www.academia.edu/9606097/La_lengua_de_palenque_avances_en_la_investigaci%C3%B3n_de_su_estructura_gramatical',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Palenquero Creole: the syntax of 2nd person pronouns and the pragmatics of address switching',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.png',
    url: 'https://www.academia.edu/34881344/Palenquero_Creole_the_syntax_of_2nd_person_pronouns_and_the_pragmatics_of_address_switching',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Diccionario de la lengua afro palenquera - español',
    image: 'https://books.google.com.co/books/content?id=znoCDR1nr0QC&printsec=frontcover&img=1&zoom=5&edge=curl&imgtk=AFLRE73vsu7MDz4lUcy2H_ssyd5aM-9rT68jlF0QhK-wOTZ51j0P2R6WkN2qDC7spiMv_JKh0CzLKZ8qqPlbKtvSf8D7vlkAs2slWcA-kHBM2o3IXI_qQSA2OWVJYul2qHSYTDXlcTCk',
    url: 'https://books.google.com.co/books?id=znoCDR1nr0QC&printsec=frontcover#v=onepage&q&f=false',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Palenquero Creole Spanish',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.png',
    url: 'http://www.personal.psu.edu/jml34/Palenque-part.pdf',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'PALENQUE DE SAN BASILIO Obra Maestra del Patrimonio Intangible de la Humanidad',
    image: 'https://docplayer.es/docs-images/46/12380028/images/page_1.jpg',
    url: 'https://www.urosario.edu.co/Subsitio/Catedra-de-Estudios-Afrocolombianos/Documentos/03-Presentacion-Dossier-Unesco---Palenque-de-San-B.pdf',
    type: BookCategoryEnum.CULTURA
  },
  {
    name: 'La época de formación de la lengua de Palenque: datos históricos y lingüísticos',
    image: 'https://0.academia-photos.com/attachment_thumbnails/37246852/mini_magick20190304-12149-15k2bgm.png?1551710463',
    url: 'https://www.academia.edu/11850884/La_%C3%A9poca_de_formaci%C3%B3n_de_la_lengua_de_Palenque_datos_hist%C3%B3ricos_y_ling%C3%BC%C3%ADsticos',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'LES CONSTRUCTIONS GÉNITIVES EN PALENQUERO : UNE SÉMANTAXE AFRICAINE',
    image: 'https://0.academia-photos.com/attachment_thumbnails/33132474/mini_magick20190406-17423-19f70u0.png?1554535643',
    url: 'https://www.academia.edu/6284935/LES_CONSTRUCTIONS_G%C3%89NITIVES_EN_PALENQUERO_UNE_S%C3%89MANTAXE_AFRICAINE',
    type: BookCategoryEnum.LENGUA
  }, {
    name: 'Oralitura de San Basilio de Palenque: temas europeos, africanos y criollos',
    image: 'https://0.academia-photos.com/attachment_thumbnails/38201141/mini_magick20190226-20016-1gbi7dj.png?1551175076',
    url: 'https://www.academia.edu/14091210/Graciela_Maglia_e_Yves_Mo%C3%B1ino_Oralitura_de_San_Basilio_de_Palenque_temas_europeos_africanos_y_criollos_',
    type: BookCategoryEnum.PALENQUE
  }, {
    name: 'Tambien existiamos en aquellos tiempos coloniales',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.png',
    url: 'https://www.academia.edu/42203874/Tambien_existiamos_en_aquellos_tiempos_coloniales',
    type: BookCategoryEnum.PALENQUE
  }, {
    name: 'AFRODESCENDIENTES EN COLOMBIA:COMPILACIÓN BIBLIOGRÁFICA',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.png',
    url: 'http://www.ram-wan.net/restrepo/documentos/afro-bibliografia-2008.pdf',
    type: BookCategoryEnum.PALENQUE
  },
  {
    name: 'Prácticas religiosas de negros',
    image: 'https://0.academia-photos.com/attachment_thumbnails/54781160/mini_magick20190115-7562-jquo0v.png?1547622043',
    url: 'https://www.academia.edu/34919895/Pr%C3%A1cticas_religiosas_de_negros_pdf',
    type: BookCategoryEnum.RELIGION
  },
  {
    name: 'Biblioteca de literatura Afrocolombiana',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.png',
    url: 'http://babel.banrepcultural.org/cdm/search/collection/p17054coll7',
    type: BookCategoryEnum.LITERATURA
  },
  {
    name: 'El Espacio cultural de Palenque de San Basilio',
    image: 'https://i.ibb.co/N7XNZx6/Captura-de-pantalla-de-2020-07-20-22-21-36.pn',
    url: 'https://unesdoc.unesco.org/ark:/48223/pf0000212738?posInSet=1&queryId=abec1e6b-0006-4737-9c8d-983ce506cd2c',
    type: BookCategoryEnum.PALENQUE
  },
  {
    name: 'Le Rôle de la langue palenge dans la transmission du patrimoine afro-palenquero',
    image: 'https://unesdoc.unesco.org/in/rest/Thumb/image?id=p%3A%3Ausmarcdef_0000162986_fre&author=P%C3%A9rez+Tejedor%2C+Juana+Pabla&title=Le+R%C3%B4le+de+la+langue+palenge+dans+la+transmission+du+patrimoine+afro-palenquero&year=2008&TypeOfDocument=UnescoPhysicalDocument&mat=ART&ct=true&size=256&isPhysical=1',
    url: 'https://unesdoc.unesco.org/ark:/48223/pf0000162986_fre?posInSet=3&queryId=abec1e6b-0006-4737-9c8d-983ce506cd2c',
    type: BookCategoryEnum.LENGUA
  },
  {
    name: 'Katalina Luango: mediadora para los difuntos entre el aquí y el allá',
    image: 'https://unesdoc.unesco.org/in/rest/Thumb/image?id=p%3A%3Ausmarcdef_0000212737&author=Simarra+Obeso%2C+Rutsely&title=Katalina+Luango%3A+mediadora+para+los+difuntos+entre+el+aqu%C3%AD+y+el+all%C3%A1&year=2009&publisher=UNESCO+Office+Havana&TypeOfDocument=UnescoPhysicalDocument&mat=ART&ct=true&size=256&isPhysical=1',
    url: 'https://unesdoc.unesco.org/ark:/48223/pf0000212737?posInSet=2&queryId=abec1e6b-0006-4737-9c8d-983ce506cd2c',
    type: BookCategoryEnum.TRADICION_ORAL
  },
  {
    name: 'Significado del Lumbalú, ritual funerario del Palenque de San Basilio',
    image: 'https://www.uninorte.edu.co/documents/7399101/14868499/Huellas+103+.jpg/b154ae91-73dd-4cb8-911a-cc44f76e745d?t=1571169978319',
    url: 'http://manglar.uninorte.edu.co/calamari/bitstream/handle/10738/72/BDC341.pdf%20lumbalu.pdf?sequence=3',
    type: BookCategoryEnum.RELIGION
  },
  {
    name: 'Cocina palenquera para el mundo',
    image: 'https://unesdoc.unesco.org/in/rest/Thumb/image?id=p%3A%3Ausmarcdef_0000245015&isbn=9789588496443&author=Ardila+Cuesta%2C+Rodolfo&title=%28Cocina+palenquera+para+el+mundo%29&year=2014&publisher=Fundaci%C3%B3n+para+el+Desarrollo+Social&TypeOfDocument=UnescoPhysicalDocument&mat=BKS&ct=true&size=256&isPhysical=1',
    url: 'https://unesdoc.unesco.org/ark:/48223/pf0000245015?posInSet=3&queryId=81b9d854-c2d4-4b36-a713-3e9184774e4e',
    type: BookCategoryEnum.GASTRONOMIA
  }
]

export default BOOKS;